Move types.ts to shared/types/colors.ts
